requirements.txt
requests
beautifulsoup4
python_dotenv
openai
asyncio
pypdf2
jsonschema
aiohttp